Added Jeff Slarve's Datafier global data manager
Along with Datafier there's a non-ABC implementation of Clarion's
critical section object called DCL_System_Threading_CriticalSection. The
Datafier uses a sync object, and the code has been updated to use the
DCL version. This is to keep the DCL library free of any ABC
dependencies. As well, an instance of the Datafier is exported from the
lib as gDatafier.